Update Houses_Sales_Analysis
The provided hyperlink will direct you to a Tableau Public platform where I conducted a thorough analysis of residential properties, honing in on room counts and monthly sales data for the years 2014 and 2015. To begin, I created a detailed line chart to visually represent the average sales of houses over the specified period. Following that, I crafted a map, utilizing a street background to illustrate the corresponding average prices in a spatial context.

Expanding the analysis further, I developed a histogram to provide insights into the distribution of house prices, offering a comprehensive view of the range and frequency of different price points. This analytical approach was then duplicated to examine the distribution of bedrooms and bathrooms, providing a more detailed breakdown of these crucial property features.

For a holistic assessment of the properties, I introduced a heatmap to compare the view with the overall conditions of the house. This allowed for a nuanced understanding of how various factors might influence property values.

To facilitate user interaction and exploration, a comprehensive calendar was meticulously constructed, incorporating a variety of filters. This enhanced the flexibility and depth of the analysis, allowing users to tailor their exploration based on specific criteria and timeframes.

In summary, I concluded the analysis by consolidating all these features into an interactive dashboard. This dashboard serves as a centralized hub, enabling users to seamlessly navigate and obtain a day-by-day analysis with just a simple click. It provides a powerful tool for gaining insights into the dynamics of residential properties, offering a multidimensional perspective on room counts, sales trends, and various property features.